refactor(ui): split PrivacyDashboardScreen into section widgets
PrivacyDashboardScreen was 491 lines and held every section widget
inline. Move them to lib/features/profile/presentation/widgets/
privacy_dashboard/: PrivacyBanner, LocalDataCard, SyncedDataCard,
PrivacyDataRow, and the three action buttons. The screen itself
drops to 152 lines and only holds the scaffold, the three callback
handlers and the delete confirmation dialog.

Closes #392
