+++ lib/features/profile/presentation/widgets/privacy_dashboard/privacy_banner.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+
+import '../../../../../l10n/app_localizations.dart';
+
+/// Top banner on the Privacy Dashboard — reassures the user that the app
+/// respects data ownership.
+class PrivacyBanner extends StatelessWidget {
+  const PrivacyBanner({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final l = AppLocalizations.of(context);
+    return Container(
+      padding: const EdgeInsets.all(16),
+      decoration: BoxDecoration(
+        color: theme.colorScheme.primaryContainer.withValues(alpha: 0.3),
+        borderRadius: BorderRadius.circular(12),
+      ),
+      child: Row(
+        children: [
+          Icon(Icons.shield, color: theme.colorScheme.primary, size: 32),
+          const SizedBox(width: 16),
+          Expanded(
+            child: Text(
+              l?.privacyDashboardBanner ??
+                  'Your data belongs to you. Here you can see everything this app stores, export it, or delete it.',
+              style: theme.textTheme.bodyMedium?.copyWith(
+                color: theme.colorScheme.primary,
+                fontWeight: FontWeight.w500,
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
