+++ lib/features/profile/presentation/widgets/privacy_dashboard/privacy_data_row.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+
+/// Small icon + label + value row used by the privacy dashboard cards.
+class PrivacyDataRow extends StatelessWidget {
+  final IconData icon;
+  final String label;
+  final String value;
+
+  const PrivacyDataRow({
+    super.key,
+    required this.icon,
+    required this.label,
+    required this.value,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 4),
+      child: Row(
+        children: [
+          Icon(icon, size: 18, color: theme.colorScheme.onSurfaceVariant),
+          const SizedBox(width: 12),
+          Expanded(
+            child: Text(label, style: theme.textTheme.bodyMedium),
+          ),
+          Text(
+            value,
+            style: theme.textTheme.bodyMedium?.copyWith(
+              fontWeight: FontWeight.w600,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
